analysis/shared/analyzers/ghidra_manager.py
import os, jpype, pyghidra, time 
from pathlib import Path
pyghidra.start() 

class GhidraManager:
    def __init__(self):
        self.launcher = None
        self.current_program = None
    
    def analyze_binary(self, binary_path, project_location=None, project_name=None):
        """
        Analyze a binary with Ghidra using PyGhidra. 
    
        Args: 
            binary_path (str): Path to the binary file to analyze.
            project_location: Where to store Ghidra Project (default: /tmp).
            project_name: Name of the Ghidra Project (default: auto-generated).

        Returns: 
            Analyzed program object.
        """

        binary_path = Path(binary_path)

    def get_functions(self):
        """Get all functions from current program."""

    def decompile_function(self, function_address):
        """
        Decompile a function at the given address.
        
        Args:
            function_address (str): Address of the function to decompile.

        Returns:
            Decompiled code as a string.
        
        """
    def get_cfg(self, function_address):
        """
        Get the control flow graph (CFG) of a function at the given address.
        
        Args:
            function_address (str): Address of the function.

        Returns:
            Dict with nodes and edges for visualization.
        """
    def get_strings(self, min_length=4):
        """
        Extract strings from the current program.
        
        Args:
            min_length (int): Minimum length of strings to extract. Default is 4.
        Returns:
            List of extracted strings.
        """
    def get_imports(self):
        """Get imported functions from the current program."""

    def rename_function(self, function_address, new_name):
        """
        Rename a function at the given address.
        """

    def add_comment(self, function_address, comment):
        """
        Add a comment to a function at the given address.
        """
    def close(self):
        """Close the Ghidra session."""
        if self.curent_program:
            self.current_program.release(self)
            self.current_program = None

        print("Ghidra session has ended.")
     
if __name__ == "__main__":
    