requirements.txt
altgraph==0.17.5
bcrypt==5.0.0
capstone==6.0.0a5
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colored-traceback==0.4.2
cryptography==46.0.3
idna==3.11
intervaltree==3.2.1
invoke==2.2.1
jfx-bridge==1.0.0
jpype1==1.5.2
macholib==1.16.4
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
paramiko==4.0.0
pefile==2024.8.26
plumbum==1.10.0
psutil==7.2.1
pwntools==4.15.0
pycparser==2.23
pyelftools==0.32
pyghidra==3.0.0
Pygments==2.19.2
PyNaCl==1.6.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.5
ROPGadget==7.7
rpyc==6.0.2
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
unicorn==2.1.2
unix-ar==0.2.1
urllib3==2.6.2
websockets==15.0.1
zstandard==0.25.0
